output-mapping: PK helpers moved from TableStructureModifier to AbstractTableStructureModifier
